requirements.txt
google_api_python_client==2.86.0
google_auth_oauthlib==1.0.0
praw==7.8.1
protobuf==3.20.3
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.22.0
tweepy==4.14.0
